reorganize linters to allow individual enable/disable
[ci skip-rust]

[ci skip-build-wheels]
